test: 인수테스트 격리를 위한 AcceptanceTest 클래스 생성
