+++ src/test/java/nextstep/subway/station/StationAcceptanceTest.java
@@ -3,6 +3,7 @@ package nextstep.subway.station;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import nextstep.subway.AcceptanceTest;
 import nextstep.subway.testutils.DatabaseCleanup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -20,26 +21,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
 @DisplayName("지하철역 관련 기능")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("acceptance")
-public class StationAcceptanceTest {
+public class StationAcceptanceTest extends AcceptanceTest {
     public static final String HEADER_LOCATION = "Location";
-    public static final String KEY_STATION_NAME = "name";
-
-    @LocalServerPort
-    int port;
-
-    @Autowired
-    private DatabaseCleanup databaseCleanup;
-
-    @BeforeEach
-    public void setUp() {
-        if (RestAssured.port == RestAssured.UNDEFINED_PORT) {
-            RestAssured.port = port;
-            databaseCleanup.afterPropertiesSet();
-        }
-        databaseCleanup.execute();
-    }
 
     /**
      * When 지하철역을 생성하면
@@ -123,6 +106,6 @@ public class StationAcceptanceTest {
     }
 
     private List<String> stationNames(ExtractableResponse<Response> response) {
-        return response.jsonPath().getList(KEY_STATION_NAME, String.class);
+        return response.jsonPath().getList("name", String.class);
     }
 }
